IMPALA-13882: Fix Iceberg v2 deletes with tuple caching
A variety of Iceberg statements (including v2 deletes) rely
on getting information from the scan node child of the
delete node. Since tuple caching can insert a TupleCacheNode
above that scan, the logic is currently failing, because
it doesn't know how to bypass the TupleCacheNode and get
to the scan node below.

This modifies the logic in multiple places to detect a
TupleCacheNode and go past it to the get the scan node
below it.

Testing:
 - Added a basic Iceberg test with v2 deletes for the
  frontend test and custom cluster test

Change-Id: I162e738c4e4449a536701a740272aaac56ce8fd8
Reviewed-on: http://gerrit.cloudera.org:8080/22666
Reviewed-by: Kurt Deschler <kdeschle@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>